+++ src/main.rs
@@ -1,9 +1,10 @@
 use actix_files::Files;
 use actix_multipart::form::MultipartFormConfig;
 use actix_multipart::form::{MultipartForm, tempfile::TempFile};
-use actix_web::HttpResponse;
 use actix_web::Responder;
+use actix_web::middleware::from_fn;
 use actix_web::{self, App, HttpServer, get, post};
+use actix_web::{HttpResponse, web};
 use askama::Template;
 use clap::Parser;
 use local_ip_address::local_ip;
@@ -17,19 +18,20 @@ use webbrowser::open;
 
 mod cli;
 mod routes;
-mod templates;
-mod utils;
 use cli::Args;
+use routes::authentication::authentication;
 use routes::download::download_page;
 use routes::index::index;
+use routes::login::login;
 use routes::upload::upload_page;
 use routes::upload_files::upload;
-use templates::download::DownloadTemplate;
-use templates::index::IndexTemplate;
-use templates::upload::UploadTemplate;
-use utils::store::copy_files_to_temp;
-use utils::types::FileInfo;
-use utils::types::UploadForm;
+mod store;
+use store::copy_files_to_temp;
+mod types;
+use types::FileInfo;
+use types::UploadForm;
+mod middleware;
+use middleware::check_auth;
 
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
@@ -39,6 +41,7 @@ async fn main() -> std::io::Result<()> {
     let port = args.get_port();
     let files = args.get_files();
     let nuke = args.get_nuke();
+    let auth = web::Data::new(args.get_auth());
 
     if nuke && temp_dir.exists() {
         fs::remove_dir_all(temp_dir)?;
@@ -76,6 +79,10 @@ async fn main() -> std::io::Result<()> {
             .service(upload_page)
             .service(download_page)
             .service(upload)
+            .service(login)
+            .service(authentication)
+            .wrap(from_fn(check_auth))
+            .app_data(auth.clone())
             .app_data(
                 MultipartFormConfig::default()
                     .total_limit(10 * 1024 * 1024 * 1024) // 10 GB