+++ src/templates.rs
@@ -1,3 +1,0 @@
-pub mod download;
-pub mod index;
-pub mod upload;